feat(project): Add ability to package the extension pack
Now that we have the basics, we need a means to package up the
extension. I don't want to publish it to the marketplace cause frankly,
MS don't make it easy. It is just as easy for one dev to package up
this extension and circulate it among others, at least for now.

I added a section to the readme to explain the steps of packaging
the extension and how to install it.